requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.2
0pFlask-Cors==4.0.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
PyJWT==2.9.0
setuptools==68.2.2
SQLAlchemy==2.0.29
typing_extensions==4.10.0
Werkzeug==3.0.1
wheel==0.41.2
pip install blinker==1.7.0
pip install click==8.1.7
pip install colorama==0.4.6
pip install Flask==3.0.2
pip install Flask-Cors==4.0.0
pip install Flask-JWT-Extended==4.7.1
pip install Flask-SQLAlchemy==3.1.1
pip install greenlet==3.1.1
pip install itsdangerous==2.1.2
pip install Jinja2==3.1.3
pip install MarkupSafe==2.1.5
pip install mysql-connector-python==8.3.0
pip install PyJWT==2.9.0
pip install setuptools==68.2.2
pip install SQLAlchemy==2.0.29
pip install typing_extensions==4.10.0
pip install Werkzeug==3.0.1
pip install wheel==0.41.2
